Rename `doc-store*` -> `content-store`.
This disambiguates from the term `doc` as used at a higher layer.
