+++ local-modules/content-store-local/main.js
