+++ local-modules/content-store/BaseDoc.js
