+++ local-modules/content-store/BaseDocStore.js
