+++ local-modules/content-store/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "doc-store",
+  "name": "content-store",
   "version": "1.0.0",
   "main": "main.js",
