+++ local-modules/doc-server/package.json
@@ -7,8 +7,8 @@
     "weak": "^1.0.1",
 
     "api-common": "local",
+    "content-store": "local",
     "doc-common": "local",
-    "doc-store": "local",
     "hooks-server": "local",
     "see-all": "local",
     "server-env": "local",
